自定义的mlp

+++ core/layer/fused_mlp.py
@@ -66,7 +66,7 @@ def gemm_kernel(
         a_ptrs += BLOCK_SIZE_K * stride_ak
         b_ptrs += BLOCK_SIZE_K * stride_bk
 
-    c = accumulator.to(a.dtype)  # 动态匹配输入类型
+    c = accumulator.to(torch.bfloat16)  # 动态匹配输入类型
     offs_cm = pid_m * BLOCK_SIZE_M + tl.arange(0, BLOCK_SIZE_M)
     offs_cn = pid_n * BLOCK_SIZE_N + tl.arange(0, BLOCK_SIZE_N)
     c_ptrs = c_ptr + stride_cm * offs_cm[:, None] + stride_cn * offs_cn[None, :]
